fix: poll for publish completion in interactive sdk publish
Interactive `sdk publish` returned as soon as the publishing API accepted the
request, printing a link and exiting 0 regardless of the outcome, while
non-interactive polled to a terminal state and exited non-zero on failure.

Move the polling, the running notice, and the closing log-URL note into the
shared `SdkPublishAction`, so both modes report the same outcome and exit code.
`pollPublishingStatus`, `publishingRunningNotice`, and `postPublishingMessage`
move to the shared `SdkPublishPrompts`; the interactive-only
`sdkPublishingInProgress` note is dropped since publishing is no longer in
flight when the command exits.

The poller now reports `succeeded | failed | cancelled` instead of a boolean.
The spinner prints its own cancel line and removes its signal listeners on
Ctrl+C, so the previous loop kept polling invisibly until a second Ctrl+C
killed the process; it now aborts the wait, tells the user publishing continues
on APIMatic, prints the log URL, and exits 130.

Refs #312, #313
